requirements.txt
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycodestyle==2.7.0
PyJWT==1.7.1
python-dotenv==0.16.0
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
twilio==6.55.0
urllib3==1.26.4
